# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: M360 API
  version: 1.0.0
  description: |
    M360 SME Lending CRM API for Shariah-compliant Murabaha financing in Saudi Arabia.
    Bilingual Arabic/English with full loan lifecycle management, Saudi government
    integrations (SIMAH, Bayan, Nafath, Yaqeen, Watheq), and role-based access control.

    ## Roles
    - `admin` - Full system access
    - `manager` - Management-level access
    - `loan_officer` - Loan origination and management
    - `credit_analyst` - Credit assessment and committee packages
    - `compliance_officer` - Compliance oversight
    - `collections_officer` - Collections management
    - `data_entry` - Data entry operations

    ## Application Workflow
    Draft -> Submitted -> PreApproved -> DocsCollected -> CreditAssessment -> CommitteeReview -> Approved/Rejected -> Disbursed
  contact:
    name: M360 Team
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - BearerAuth: []
...
paths:
...
  /organizations:
    post:
      tags: [Organizations]
      summary: Create an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreateRequest'
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Organizations]
      summary: List organizations
      parameters:
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Paginated list of organizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /organizations/{id}:
    get:
      tags: [Organizations]
      summary: Get organization by ID
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Organization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Organizations]
      summary: Update an organization
      parameters:
        - $ref: '#/components/parameters/IDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdateRequest'
      responses:
        '200':
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Organizations]
      summary: Delete an organization
      description: Requires admin or manager role.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '204':
          description: Organization deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Contacts
  # ──────────────────────────────────────────────
...
  /contacts/organization/{org_id}:
    get:
      tags: [Contacts]
      summary: List contacts by organization
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of contacts for the organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Leads
  # ──────────────────────────────────────────────
...
  /products:
    post:
      tags: [Products]
      summary: Create a product
      description: Requires admin or manager role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Products]
      summary: List products
      parameters:
        - name: active
          in: query
          description: Filter to active products only
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Products]
      summary: Update a product
      description: Requires admin or manager role.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Applications
  # ──────────────────────────────────────────────
...
  /reports/dashboard:
    get:
      tags: [Reports]
      summary: Get dashboard statistics
      description: Overall statistics including leads, applications, facilities, disbursed amounts, and PAR metrics.
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports/pipeline:
    get:
      tags: [Reports]
      summary: Get pipeline statistics
      description: Application pipeline funnel showing count per workflow status.
      responses:
        '200':
          description: Pipeline statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports/officers:
    get:
      tags: [Reports]
      summary: Get officer performance
      description: Performance metrics for loan officers. Requires manager or admin role.
      responses:
        '200':
          description: Officer performance list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OfficerPerformance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports/par:
    get:
      tags: [Reports]
      summary: Get portfolio at risk analysis
      description: Delinquency breakdown and portfolio at risk ratios. Requires manager or admin role.
      responses:
        '200':
          description: Portfolio at risk data
          content:
            application/json:
              schema:
                type: object
                description: Portfolio at risk analysis with delinquency breakdown
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/refresh
  parameters:
    IDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SearchParam:
      name: search
      in: query
      description: Search term
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      description: Number of items per page (default 20)
      schema:
        type: integer
        default: 20
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name_en:
          type: string
        name_ar:
          type: string
        cr_number:
          type: string
          nullable: true
        cr_verified:
          type: boolean
        tax_id:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        legal_structure:
          type: string
          nullable: true
        founding_date:
          type: string
          format: date-time
          nullable: true
        address_en:
          type: string
          nullable: true
        address_ar:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        assigned_officer_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrganizationCreateRequest:
      type: object
      required: [name_en]
      properties:
        name_en:
          type: string
        name_ar:
          type: string
        cr_number:
          type: string
        tax_id:
          type: string
        industry:
          type: string
        legal_structure:
          type: string
        address_en:
          type: string
        address_ar:
          type: string
        city:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
    OrganizationUpdateRequest:
      type: object
      properties:
        name_en:
          type: string
        name_ar:
          type: string
        cr_number:
          type: string
        tax_id:
          type: string
        industry:
          type: string
        legal_structure:
          type: string
        address_en:
          type: string
        address_ar:
          type: string
        city:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
    OrganizationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        total:
          type: integer

    # ── Contact ──
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name_en:
          type: string
        name_ar:
          type: string
        national_id:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        nafath_verified:
          type: boolean
        simah_score:
          type: integer
          nullable: true
        is_guarantor:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name_en:
          type: string
        name_ar:
          type: string
        type:
          type: string
        min_amount:
          type: number
          format: double
        max_amount:
          type: number
          format: double
        min_tenor_months:
          type: integer
        max_tenor_months:
          type: integer
        profit_rate:
          type: number
          format: double
        admin_fee_pct:
          type: number
          format: double
        payment_frequency:
          type: string
        eligibility_criteria:
          type: object
          description: Arbitrary JSON
        required_documents:
          type: object
          description: Arbitrary JSON
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductCreateRequest:
      type: object
      required: [name_en]
      properties:
        name_en:
          type: string
        name_ar:
          type: string
        type:
          type: string
        min_amount:
          type: number
          format: double
        max_amount:
          type: number
          format: double
        min_tenor_months:
          type: integer
        max_tenor_months:
          type: integer
        profit_rate:
          type: number
          format: double
        admin_fee_pct:
          type: number
          format: double
        payment_frequency:
          type: string
        required_documents:
          type: object
    ProductUpdateRequest:
      type: object
      properties:
        name_en:
          type: string
        name_ar:
          type: string
        min_amount:
          type: number
          format: double
        max_amount:
          type: number
          format: double
        profit_rate:
          type: number
          format: double
        admin_fee_pct:
          type: number
          format: double
        is_active:
          type: boolean

    # ── Application ──
...
    DashboardStats:
      type: object
      properties:
        total_leads:
          type: integer
        total_applications:
          type: integer
        total_facilities:
          type: integer
        total_disbursed:
          type: number
          format: double
        total_outstanding:
          type: number
          format: double
        par_30:
          type: number
          format: double
        par_60:
          type: number
          format: double
        par_90:
          type: number
          format: double
    PipelineStats:
      type: object
      properties:
        draft:
          type: integer
        submitted:
          type: integer
        pre_approved:
          type: integer
        documents_collected:
          type: integer
        credit_assessment:
          type: integer
        committee_review:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
        disbursed:
          type: integer
    OfficerPerformance:
      type: object
      properties:
        officer_id:
          type: string
        officer_name:
          type: string
        lead_count:
          type: integer
        app_count:
          type: integer
        disbursed:
          type: number
          format: double
        conversion_rate:
          type: number
          format: double